+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - General
   - Fix classpath scan when classpath has other things like new lines or warning message besides the classpath. Fixes #523
   - Improve `clean-ns` to remove empty reader conditionals(`#?(:clj)`) after cleaning requires/imports.
+  - Fix `clean-ns` false-positives removals to cljc files when the alias/refer/import is being used inside a reader conditional.
 
 ## 2021.08.16-19.02.30